Add rejection-message feature
Now if inspectors input 'message' when they inspect the registration
``message`` context will be set in each email template thus the message
can be accessed as "{{ message }}" in each email template.

It is really useful to explain why the registration was rejected in
email. In default template, the message is used for rejection reason.